Clock refactor
