+++ core/shared/src/main/scala/zio/clock/Service.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2017-2020 John A. De Goes and the ZIO Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package zio.clock
+
+import java.time.OffsetDateTime
+import java.util.concurrent.TimeUnit
+
+import zio._
+import zio.duration._
+
+trait Service {
+  def currentTime(unit: TimeUnit): UIO[Long]
+  def currentDateTime: UIO[OffsetDateTime]
+  def nanoTime: UIO[Long]
+  def sleep(duration: Duration): UIO[Unit]
+}
